Refactor http payload mocking in unit tests
