+++ OutlookMatters.Core.Test/Mattermost/v1/HttpRestServiceTest.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
 using FluentAssertions;
 using Moq;
 using NUnit.Framework;
@@ -8,6 +7,7 @@ using OutlookMatters.Core.Http;
 using OutlookMatters.Core.Mattermost.v1;
 using OutlookMatters.Core.Mattermost.v1.Interface;
 using Test.OutlookMatters.Core.Http;
+using Test.OutlookMatters.Core.Mattermost.v1.Interface;
 
 namespace Test.OutlookMatters.Core.Mattermost.v1
 {
@@ -330,54 +330,6 @@ namespace Test.OutlookMatters.Core.Mattermost.v1
                 Password = USER_PASSWORD
             };
             return login;
-        }
-    }
-
-    public static class TestHelper
-    {
-        public static string SerializeToPayload(this Login login)
-        {
-            return "{\"name\":\"" + login.Name + "\",\"email\":\"" + login.Email + "\",\"password\":\"" + login.Password +
-                   "\"}";
-        }
-
-        public static string SerializeToPayload(this User user)
-        {
-            return "{\"id\":\"" + user.Id + "\"}";
-        }
-
-        public static string SerializeToPayload(this Error error)
-        {
-            return "{\"message\":\"" + error.message + "\",\"detailed_error\":\"" + error.detailed_error + "\"}";
-        }
-
-        public static string SerializeToPayload(this Post post)
-        {
-            return "{\"id\":\"" + post.Id + "\",\"channel_id\":\"" + post.ChannelId + "\",\"message\":\"" + post.Message +
-                   "\",\"user_id\":\"" + post.UserId + "\",\"root_id\":\"" + post.RootId + "\"}";
-        }
-
-        public static string SerializeToPayload(this Channel channel)
-        {
-            return "{\"id\":\"" + channel.ChannelId +
-                   "\",\"create_at\":1458911668852,\"update_at\":1458911668852,\"delete_at\":0,\"type\":\"" +
-                   channel.Type +
-                   "\",\"display_name\":\"" + channel.ChannelName + "\"}";
-        }
-
-        public static string SerializeToPayload(this Thread post)
-        {
-            return
-                "{\"order\":[\"" + string.Join(",", post.Order) + "\"],\"posts\":{" +
-                string.Join(",", post.Posts.Select(x => "\"" + x.Key + "\":" + SerializeToPayload((Post) x.Value))) +
-                "}}";
-        }
-
-        public static string SerializeToPayload(this ChannelList channelList)
-        {
-            return
-                "{\"channels\":[" + string.Join(",", channelList.Channels.Select(x => x.SerializeToPayload())) +
-                "]}";
         }
     }
 }
